src/requirements.txt
python-dotenv==0.11.0
Flask==1.1.1
Flask-RESTful==0.3.8